requirements.txt
asgiref==3.3.1
Django==3.1.5
django-3-jet==1.0.8
django-auth-ldap==2.2.0
django-ckeditor==6.1.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
Pillow==8.2.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.4
python-ldap==3.3.1
pytz==2020.5
six==1.16.0
sqlparse==0.4.1
